move sqllogictest to a file under tests
