--- tests/sqllogictest/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "risinglight_sqllogictest"
-version = "0.1.0"
-edition = "2021"
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-async-trait = "0.1"
-env_logger = "0.10"
-glob = "0.3"
-libtest-mimic = "0.5"
-risinglight = { path = "../.." }
-sqllogictest = "0.13.0"
-tempfile = "3"
-tokio = { version = "1", features = ["rt"] }
-
-[[test]]
-name = "sqllogictest"
-harness = false
